Spring Mvc,Crud,Security basics covered
This demo application demonstrates fundamental concepts of MVC, CRUD operations, and security. It is developed using Spring Boot and Spring MVC. It uses a simple database for CRUD operations, integrated with Spring Data JPA. Furthermore, user authentication and authorization are implemented using Spring Security. The application is designed to provide a basic understanding of these concepts.